+++ server/src/main/kotlin/org/kryptonmc/krypton/extension/utils.kt
@@ -3,7 +3,7 @@ package org.kryptonmc.krypton.extension
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import java.net.InetAddress
-import kotlin.math.log
+import kotlin.math.log2
 import kotlin.math.max
 
 fun String.toInetAddress(): InetAddress = InetAddress.getByName(this)
@@ -12,7 +12,7 @@ inline fun <reified T> logger(): Logger = LogManager.getLogger(T::class.java)
 
 fun logger(name: String): Logger = LogManager.getLogger(name)
 
-fun Int.calculateBits() = max(log(takeHighestOneBit().toDouble(), 2.0).toInt() + 1, 4)
+fun Int.calculateBits() = max(log2(takeHighestOneBit().toDouble()).toInt() + 1, 4)
 
 fun Int.toArea() = (this * 2 + 1).square()