+++ server/src/main/kotlin/org/kryptonmc/krypton/packet/in/play/PacketInPluginMessage.kt
@@ -1,11 +1,11 @@
 package org.kryptonmc.krypton.packet.`in`.play
 
 import io.netty.buffer.ByteBuf
-import org.kryptonmc.krypton.extension.readString
-import org.kryptonmc.krypton.extension.readVarIntByteArray
-import org.kryptonmc.krypton.packet.state.PlayPacket
 import org.kryptonmc.krypton.api.registry.NamespacedKey
 import org.kryptonmc.krypton.api.registry.toNamespacedKey
+import org.kryptonmc.krypton.extension.readAllAvailableBytes
+import org.kryptonmc.krypton.extension.readString
+import org.kryptonmc.krypton.packet.state.PlayPacket
 
 class PacketInPluginMessage : PlayPacket(0x0B) {
 
@@ -17,6 +17,6 @@ class PacketInPluginMessage : PlayPacket(0x0B) {
 
     override fun read(buf: ByteBuf) {
         channel = buf.readString().toNamespacedKey()
-        data = buf.readVarIntByteArray()
+        data = buf.readAllAvailableBytes()
     }
 }