+++ server/src/main/kotlin/org/kryptonmc/krypton/packet/out/login/PacketOutLoginDisconnect.kt
@@ -5,7 +5,7 @@ import net.kyori.adventure.text.Component
 import org.kryptonmc.krypton.extension.writeChat
 import org.kryptonmc.krypton.packet.state.LoginPacket
 
-class PacketOutDisconnect(private val component: Component) : LoginPacket(0x00) {
+class PacketOutLoginDisconnect(private val component: Component) : LoginPacket(0x00) {
 
     override fun write(buf: ByteBuf) {
         buf.writeChat(component)