+++ server/src/main/kotlin/org/kryptonmc/krypton/session/Session.kt
@@ -1,10 +1,13 @@
 package org.kryptonmc.krypton.session
 
 import io.netty.channel.Channel
+import net.kyori.adventure.text.Component
 import org.kryptonmc.krypton.auth.GameProfile
 import org.kryptonmc.krypton.entity.entities.KryptonPlayer
 import org.kryptonmc.krypton.packet.Packet
 import org.kryptonmc.krypton.packet.data.ClientSettings
+import org.kryptonmc.krypton.packet.out.login.PacketOutLoginDisconnect
+import org.kryptonmc.krypton.packet.out.play.PacketOutDisconnect
 import org.kryptonmc.krypton.packet.state.PacketState
 import kotlin.random.Random
 
@@ -25,7 +28,12 @@ class Session(val id: Int, internal val channel: Channel) {
         channel.writeAndFlush(packet)
     }
 
-    fun disconnect() {
+    fun disconnect(reason: Component) {
+        when (currentState) {
+            PacketState.PLAY -> sendPacket(PacketOutDisconnect(reason))
+            PacketState.LOGIN -> sendPacket(PacketOutLoginDisconnect(reason))
+            else -> Unit
+        }
         if (channel.isOpen) channel.close().awaitUninterruptibly()
     }
 }